==> 1-Data-Collection.ipynb <==
# Install Python Libraries
%pip install pandas matplotlib seaborn

# Import necessary libraries
import pandas as pd
import json

with open("/Users/kimberlyali/Desktop/data science projects/Spotify Extended Streaming History/Streaming_History_Audio_2024-2025_2.json") as f:
    data1 = json.load(f)

df1 = pd.DataFrame(data1)

with open("/Users/kimberlyali/Desktop/data science projects/Spotify Extended Streaming History/Streaming_History_Audio_2025_3.json") as g:
    data2 = json.load(g)

df2 = pd.DataFrame(data2)

spotify_df = pd.concat([df1, df2], ignore_index=True)
spotify_df.head()

In this table, we notice that our dataframe starts in March 2024, but we will be working from Jan 2025, giving us 7 months of data.

The table below defines each header (provided by Spotify as a PDF file)

| Technical Field | Contains |
| --- | --- |
| ts | a timestamp when the track stopped playing. The order is year, month, and day, followed by a timestamp in military time |
| username | my Spotify username |
| platform | the platform used when streaming the track |
| ms_played | the number of milliseconds the stream was played |
| conn_country | the country code of the country where the stream was played |
| ip_addr | the IP address logged when streaming the track |
| user_agent_decrypted | the user agent when streaming the track |
| master_metadata_track_name | the name of the track |
| master_metadata_album_artist_name | the name of the artist, band, or podcast |
| master_metadata_album_album_name | the name of the album of the track |
| spotify_track_uri | a Spotify URL, uniquely identifying the track in the form of "spotify:track:<base-62 string\>" |
| episode_name | the name of the episode of the podcast | 
| episode_show_name | the name of the podcast |
| episode_uri | a Spotify Episode URL, uniquely identifying the episode in the form of "spotify:episode:<base-62 string\>" |
| audiobook_uri | a Spotify Audiobook URL, uniquely identifying the audiobook in the form of "spotify:audiobook:\<base-62 string\>" |
| audiobook_chapter_uri | a Spotify Audiobook Chapter URL, uniquely identifying the audiobook chapter |
| audiobook_chapter_title | the name of the audiobook chapter |
| reason_start | a value telling why the track started (e.g. "trackdone") | 
| reason_end | a value telling why the track ended (e.g. "endplay") |
| shuffle | the value True or False depending on if shuffle mode was used when playing the track |
| skipped | indicates whether the user skipped to the next song |
| offline | indicates whether the track was played in offline mode (True) or not (False) |
| offline_timestamp | a timestamp of when offline mode was uses, if used |
| incognito_mode | indicates whether the track was played during a private session (True) or not (False) |

# Save as csv file to load in another notebook
spotify_df.to_csv("cleaned_spotify_data.csv", index=False)
