==> 2-Data-Preparation.ipynb <==
import pandas as pd
from datetime import datetime
import json

# Load Data Collection csv file
df = pd.read_csv("cleaned_spotify_data.csv")

df.head()

df.info()

df.describe(include='all')

df.isnull().sum()

There are 11 tracks and over 25K audiobook/podcast episodes missing.
First, let's drop the columns we don't need, and then decide what to do with the null-valued tracks.

# Removing columns relating to episodes, and the IP address.
df = df.drop(['episode_name', 'episode_show_name', 'spotify_episode_uri', 'audiobook_title', 'audiobook_uri', 'audiobook_chapter_uri', 'audiobook_chapter_title', 'ip_addr'], axis=1, errors='ignore')
df.head()

# Dropping the missing track
df.dropna(inplace=True)
df.isnull().sum()

Now, we will convert the timestamps and filter out songs played for less than 30 seconds

# Converting the timestamps
df['ts'] = pd.to_datetime(df['ts'])
df['hour'] = df['ts'].dt.hour
df['day'] = df['ts'].dt.day_name()
df['month'] = df['ts'].dt.to_period('M')

# Inspecting the duration played and filtering out any streams less than 30s
df['minutes_played'] = df['ms_played'] / (1000 * 60)
df = df[df['minutes_played'] > 0.5]

# Creating a column that combines the track name and its artist's name
df['track_artist'] = df['master_metadata_track_name'] + " - " + df['master_metadata_album_artist_name']

df.head()

Now, we need to filter our dataframe to be Jan-July 2025 rather than starting from March 2024.

wrapped_df = df[df['ts'].dt.year == 2025]
wrapped_df.head()

Lastly, we need to remove the duplicate values as for Spotify Wrapped, they likely groups plays per track but deduplicate certain edge cases (e.g., replays). And, we know I have a lot of those.

wrapped_df.duplicated().sum()
wrapped_df = wrapped_df.drop_duplicates()

wrapped_df.to_csv("prepared_spotify_data.csv", index=False)
