==> 5-Predictive-Logic.ipynb <==
In this section, we are going to be predicting my Spotify Wrapped. There are five main components of a Spotify Wrapped:
1. Top 5 Artists
2. Top 5 Songs
3. Favourite Genre
4. Most Active Listening Time
5. Total Listening Time Estimate

But, Spotify did not provide a genre column in the extended listening history file. So, we will be focusing on the others (and possibly some bonus categories).

We will be focusing on the following strategy:

**Trend Projection** - taking the cleaned and EDA'd data and project the rest of the year based on trends.

And, we will be using **Random Forest** and **k-Nearest Neighbors** to compare predictions.

%pip install -U scikit-learn

# Import relevant libraries (it's a lot omg)
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from datetime import datetime
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import StandardScaler, LabelEncoder
from sklearn.ensemble import RandomForestClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import classification_report, accuracy_score
from sklearn.inspection import permutation_importance

df = pd.read_csv("engineered_spotify_data.csv")

df.head()

### Trend Projection

# Top 5 Artists
top_artists_time = (
    df.groupby('master_metadata_album_artist_name')['ms_played']
    .sum()
    .sort_values(ascending=False)
    .head(5) / 60000
)

# Project for the full year
df['ts'] = pd.to_datetime(df['ts'])
months_so_far = df['ts'].dt.month.nunique()

# Scale up as if the year continued with the same pattern
# (There's a pattern? LOL)
scale_factor = 12 / months_so_far
top_artists_projected = (top_artists_time * scale_factor).round(2)

print("Projected Top 5 Artist by Listening Time (Full Year): ")
print(top_artists_projected)

# Top 5 Tracks
top_tracks_time = (
    df.groupby('track_artist')['ms_played']
    .sum()
    .sort_values(ascending=False)
    .head(5) / 60000
)

# Project for the full yeah
top_tracks_projected = (top_tracks_time * scale_factor).round(2)

print("Projected Top 5 Songs by Listening Time (Full Year): ")
print(top_tracks_projected)

# Most Active Listening Hour
hourly_listening = (
    df.groupby('hour')['ms_played']
    .sum()
    .sort_values(ascending=False)
)

top_hour = hourly_listening.idxmax()
minutes = round(hourly_listening.max() / 60000, 1)
print(f"My Peak Listening Hour: {top_hour}, with approximately {minutes} minutes listened")

# Total Listening Time Estimate
total_minutes = df['minutes_played'].sum()
print(f"Estimated Total Listening Time: {total_minutes:1f} minutes")

### Preprocessing

# Encode categorical target variable
le = LabelEncoder()
df['top_artist_encoded'] = le.fit_transform(df['top_artist'])

### Random Forest Classifier
It is a tree-based model that build multiple decision trees and averages their predictions.

### K-Nearest Neighbors (KNN)
It compares each new row with similar rows in the dataset.
